+++ TAC_AI/AI/CombatUtils.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TAC_AI.AI
+{
+    public class CombatUtils
+    {
+    }
+}
